requirements.test.txt
pytest==4.2.0
py-ecc==1.7.0
pysha3==1.0.2
bitcoin==1.1.42
